redo all code

+++ fs.js
@@ -1,12 +1,18 @@
 fs = require('fs');
 
 module.exports.writeFile = function (fileName, fileContent) {
-    fs.writeFile(fileName, fileContent, (err) => {
-        if (err)
-            console.log(err);
-        else
-            console.log('file:%s', fileName + ' has generated');
-    });
+    folder = fileName.substr(0,fileName.lastIndexOf('/'));
+    if (fs.existsSync(folder)) {
+        fs.writeFile(fileName, fileContent, (err) => {
+            if (err)
+                console.log(err);
+            else
+                console.log('file:%s', fileName + ' has generated');
+        });
+    }else{
+        this.writeFile(fileName, fileContent);
+    }
+    
 }
 module.exports.createFolder = function (folderLocate) {
     fs.mkdir(folderLocate, (err) => {
